+++ packages/@molt/command/src/index_.ts
@@ -2,4 +2,4 @@ export {
   createViaParameter as parameter,
   createViaParameters as parameters,
   createViaParametersExclusive as parametersExclusive,
-} from './Builder/root/constructor.js'
+} from './Builder/root/runtime.js'
