+++ packages/@molt/command/tests/line/builder.spec.ts
@@ -2,9 +2,6 @@ import { Command } from '../../src/index.js'
 import { describe, expect, it } from 'vitest'
 import { z } from 'zod'
 
-// eslint-disable-next-line
-const assertAssignable = <T>(_: T): [T] => 0 as any
-
 describe(`.create`, () => {
   it(`creates a parameters definition`, () => {
     const def = Command.parameters({ x: z.string() })
@@ -17,14 +14,5 @@ describe(`settings`, () => {
     Command.parameters({ x: z.string() })
       .settings({ description: `foobar` })
       .parse({ line: [`-x`, `foobar`] })
-  })
-})
-
-describe(`#.schema`, () => {
-  it(`is the schema used to define the parameters`, () => {
-    const def = Command.parameters({ x: z.string() })
-    assertAssignable<z.ZodRawShape>(def.schema)
-    assertAssignable<{ x: z.ZodString }>(def.schema)
-    expect(def.schema).toBeDefined()
   })
 })
