+++ packages/@molt/command/tests/line/types/string.spec.ts
@@ -1,9 +1,9 @@
 import { Command } from '../../../src/index.js'
-import { stdout } from '../../__helpers__.js'
 import type { IsExact } from 'conditional-type-checks'
 import { assert } from 'conditional-type-checks'
 import { describe, expect, it } from 'vitest'
 import { z } from 'zod'
+import { stdout } from '../../__mock__.js'
 
 // TODO use test.each
 describe(`errors`, () => {
